move layout.js and prettier formats files
