+++ src/pages/page-2.js
@@ -1,8 +1,8 @@
-import React from "react"
-import { Link } from "gatsby"
+import React from 'react'
+import { Link } from 'gatsby'
 
-import Layout from "../components/layout"
-import SEO from "../components/seo"
+import Layout from '../components/layout/layout'
+import SEO from '../components/seo'
 
 const SecondPage = () => (
   <Layout>
